# app-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Notatek",
        "description": "Dokumentacja API dla systemu notatek",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Pobieranie listy użytkowników",
                "operationId": "b3f5041876ff3c478354eeed0385a340",
                "responses": {
                    "200": {
                        "description": "Lista użytkowników"
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Dodawanie użytkownika",
                "operationId": "eb4bd5ca5c6b27a7bf567a6ea818f12d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "login",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "newuser"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "securepassword"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Użytkownik został dodany"
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    },
                    "400": {
                        "description": "Nie znaleziono roli"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Usuwanie użytkownika",
                "operationId": "7da83cb197359d23440e747c21e4a31e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID użytkownika do usunięcia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Użytkownik został usunięty"
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    },
                    "404": {
                        "description": "Użytkownik nie istnieje"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}/password": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Zmiana hasła użytkownika przez administratora",
                "operationId": "6aa8d85fb1f38f1f3c2dd17f87069970",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID użytkownika, którego hasło ma zostać zmienione",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "newsecurepassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hasło użytkownika zostało zmienione"
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    },
                    "404": {
                        "description": "Użytkownik nie istnieje"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Wpisz token w formacie (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}